+++ src/components/Input/Input.tsx
@@ -1,18 +1,28 @@
-import styled, { css } from "styled-components";
+import { FC } from "react";
+import styled from "styled-components";
 
 import getColor from "../../styles/themeHelpers/getColor";
 import getFontFamily from "../../styles/themeHelpers/getFontFamily";
 import { getBreakpoint } from "../../styles/utils/responsive";
 import { margin, MarginProps } from "../../styles/utils/spacing";
-import UnstyledInput from "../UnstyledInput";
+import Flex from "../Flex";
+import { IconName } from "../Icon";
+import UnstyledInput, { UnstyledInputProps } from "../UnstyledInput";
 
-const inputStyles = css`
+import InputIcon from "./InputIcon";
+
+type StyledInputProps = MarginProps & {
+  value: string;
+  icon?: IconName;
+};
+const StyledInput = styled(UnstyledInput)<StyledInputProps>`
+  color: ${getColor("black")};
   height: ${(props) => props.theme.input.height};
   border-radius: ${(props) => props.theme.input.borderRadius};
-  border-color: ${getColor((props) => props.palette.input.default.border)};
+  border-color: ${getColor((theme) => theme.palette.input.default.border)};
   border-width: ${(props) => props.theme.input.borderWidth};
   border-style: solid;
-  padding-left: 12px;
+  padding-left: ${(props) => (props.icon ? "40px" : "12px")};
   padding-right: 0;
   font-size: 16px;
   font-family: ${getFontFamily("ui")};
@@ -24,28 +34,41 @@ const inputStyles = css`
     font-size: 16px;
   }
 
-  &,
   ::placeholder {
     font-family: ${getFontFamily("ui")};
+    color: ${getColor((props) => props.palette.input.default.placeholder)};
+    opacity: 1;
   }
 
   :valid:not([value=""]) {
     border-color: ${getColor((props) => props.palette.input.valid.border)};
-  }
 
-  ::placeholder {
-    color: ${getColor((props) => props.palette.input.valid.placeholder)};
-    opacity: 1;
+    ::placeholder {
+      color: ${getColor((props) => props.palette.input.valid.placeholder)};
+    }
+
+    ~ ${InputIcon} {
+      color: ${getColor((theme) => theme.palette.input.valid.icon)};
+    }
   }
 
   ${margin}
+
+  ~ ${InputIcon} {
+    color: ${getColor((theme) => theme.palette.input.default.icon)};
+  }
 `;
 
-type InputProps = MarginProps & {
-  value: string;
+type InputProps = UnstyledInputProps & StyledInputProps;
+const Input: FC<InputProps> = (props) => {
+  const { icon } = props;
+
+  return (
+    <Flex alignItems="center">
+      <StyledInput {...props} />
+      {icon && <InputIcon outerWidth={40} size={20} name={icon} />}
+    </Flex>
+  );
 };
-const Input = styled(UnstyledInput)<InputProps>`
-  ${inputStyles}
-`;
 
 export default Input;