+++ src/components/UnstyledInput/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./UnstyledInput";
+export type { UnstyledInputProps } from "./UnstyledInput";